corregido error al eliminar cuenta y confirmación en forma de toast al eliminarla

+++ src/app/services/user/user.service.ts
@@ -42,6 +42,8 @@ export class UserService {
   }
 
   async removeProfilePhoto() {
+    if (!this.auth.currentUser?.photoURL)
+      return
     await updateProfile(this.auth.currentUser!, { photoURL: '' })
     await this.profilePhotoService.deleteUserProfilePhoto(this.auth.currentUser!.uid)
   }
